custo

+++ registo.c
@@ -19,7 +19,7 @@ void guarda_saida_no_registo(Registo* registo, Data* data_saida) {
     registo->saida = data_saida;
 }
 
-void guarda_custo_no_registo(Registo* registo, int custo) {
+void guarda_custo_no_registo(Registo* registo, float custo) {
     registo->custo = custo;
 }
